exceptions: improve non-error @(throw) and @(assert).
When @(throw) generates a non-error exception that is
unhandled, we just want it to continue. In the same situation,
an @(assert) should behave as a failed match; that is, the
failure of the query material that follows the assert, which
activated it, should propagate through the assert.

* match.c (v_throw): Return next_spec_k if uw_rthrow
returns.
(v_assert, h_assert): Return nil if uw_rthrow returns.

* txr.1: Expanded @(throw) and @(assert) documentation with
discussion of unhandled exceptions.