txr: spurious retention in @(next).
* match.c (mf_file_lazy): New static function. The
lazy list is created here and stored directly into
the data field of the context structure.
Function is marked NOINLINE because on an older
system with gcc 4.4.5, it didn't solve the problem.
We need the function to have a stack frame so any
spurious copies of the linked list go into a frame
that disappears when the function returns.
(v_next_impl): Use mf_file_lazy instead of mf_file_data.
(open_data_source): Also make this function INLINE
just in case because it contains calls to lazy_stream_cons.