bugfix: definitions must trigger autoload.
When a function, macro, variale, symbol macro or struct is
being defined, we must trigger any auto-load for that symbol.
If the definition is redefining a library symbol, then if
the autoload is later triggered, it will surprisingly
reinstate the library definition.

* eval.c (rt_defvarl, op_defsymacro, rt_defsymacro, rt_defun,
rt_defmacro): Insert calls to lisplib_try_load against the
symbol being defined.

* struct.c (make_struct_type): Likewise.