unicode: character width upkeep.
Updating the regex for matching code points corresponding to
wide and full width characters, with regard to the old
1998 document: http://www.unicode.org/reports/tr11-2/
More to follow.

I neglected to comment where the original data came from,
and neglected to comment it. In some cases it has more
coverage than the 1998 document; in some cases less.

* regex.c (create_wide_cs): Extending the 1100-115F range
to 11F9 to cover all of Korean Hangeul. Replace two
occurrences of 3000-303E with one 3000-303F.
Merge 3250-32FE with 3300-4DB5, and extend to 4DBF.
Add private use range E000-E757.