seq_info: bugfix: uninitialized type field.
* lib.c (seq_info): Ensure the type field in the returned
structure is inintialized. We are neglecting this when the
type is COBJ, initializing only the kind field. This makes
the in function behave unreliably over arguments that
are hash tables, or vector-like sequences (objects with
a length function or the carray type). Several other functions
will behave unreliably for vector-like sequences: reverse,
find, rfind, pos, rpos and tprint.