arith: move c_unum and unum declaration into lib.h
The c_unum and unum functions are fairly often needed
and and are closely related to c_num and num, which have
always been declared in lib.h even though they live in
arith.c. Files not doing arithmetic are including "arith.h"
just for the sake of c_num or unum.

* arith.h (c_unum, unum): Declarations removed.

* lib.h (c_unum, unum): Declarations moved here.

* chksum.c, debug.c, stream.c,  strudel.c, sysif.c, vm.c:
These files no longer require #include "arith.h" as a result,
which is therefore removed.