New style iteration functions.
iter-begin provides a paradigm for iteration that is more
compatible with lists. If the sequence is a list, then the
list itself is returned as the iterator, and the other
functions are basicaly wrappers for car/cdr and null testing.
Yet the API is defined in such a way that other objects can be
iterated with good efficiency, at the cost of allocating a new
iterator object (which can be re-used).

* eval.c (eval_init): Register iter-begin, iter-more,
iter-item, iter-step and iter-reset.

* lib.c (seq_iter_init_with_info): New static function.
(seq_iter_init): Now a thin wrapper for
seq_iter_init_with_info.
(iter_begin, iter_more, iter_item, iter_step, iter_reset): New
functions.

* lib.h (iter_begin, iter_more, iter_item, iter_step,
iter_reset): New functions.

* txr.1: Documented.