pattern lang: vertical-horizontal fallback regression.
Commit c9cab7138636c6c1d6e47f8d1a4053bec2dd0ad4, on Feb 1,
2019, breaks the following case:

  @(define horiz)whatever@(end)
  @(horiz)

The code assumes that if v_fun returned :decline, then the
function doesn't exist. That is false, because the function
may have a horizontal definition, as in the above example.
That's why in the original code, nothing was done in this case
to just allow the flow to proceed to the horizontal fallback,
where the call will be tried as a horizontal function.

The small problem with that was lack of a diagnosis when the
function actually doesn't exist (neither vertical nor
horizontal). In that case, the horizontal fallback will expect
a line of data. If no data arrives, then the undefined
function call is undiagnosed.

The idea behind c9cab7138636c6c1d6e47f8d1a4053bec2dd0ad4 was
to provide a diagnostic immediately. The right way to do that
is to check for a horizontal definition of the function. If
there isn't one, then error out, otherwise fall back on the
horizontal processing.