struct: bugfix: autoload on instance slot also.
* struct.c (slot_types): Try lisplib_try_load, just like in
static_slot_types. The reason for this is that there is code
like (or (sys:slot-types slot) (sys:static-slot-types slot))
which occurs in sys:check-slot, testing both hashes.  This
code gets fooled in the following way: the sys:slot-types slot
reports nil for a particular slot being probed. It doesn't try
autoloading, checking the hash just once. Next,
sys:static-slot-types is called and triggers an autoload. The
autoload instantiates the slot as an instance slot (but not as
a static slot).  Then sys:static-slot-types returns nil
also because there is no such static slot. The effect is that
the slot exists due to the autload, but since both functions
yielded nil, the overall (or ...) expression yields nil, which
is interpreted as the slot not existing.