uref/qref: printing bug.
* lib.c (obj_print_impl): The simple_qref_args_p function is
called with reversed values of pos. For uref, we want 1,
because the first item occurs after the dot, and for qref we
want 0, because the first item occurs before the dot.  This
mix up means that the degenerate syntax (qref X) incorrectly
prints as just X instead of (qref X). Likewise, though
harmlessly, (uref X) fails to render as .X, instead coming out
as (uref X).