uref/qref: read-print consistency issue.
The printer for uref/qref doesn't check for symbols that
contain digits in the name, and so for instance (qref a3 3a)
comes out as a3.3a, which does not read back, due to looking
like a cramped floating-point constant. It looks like the
weakest rule we can make is that all symbols that can be
preceded by a dot must not have a name starting with a digit.
Thus (qref 3a b c) can render as 3a.b.c. But (qref a 3b c)
must not render as a.3b.c; it must be (qref a 3b c).

* lib.c (simple_qref_args_p): Enact the above rule. In
all positions preceded by a dot, if the symbol starts with a
digit, then return nil to indicate that the syntax must be
rendered in the list form.