tags: address small issue with tag lookup.
Exuberant Ctags uses the full content of one line as the key
to find a tag. A function declaration that is split into
multiple lines can have a first line which is identical to the
definition, as in:

  static int add(int a,
                 int b);

  static int add(int a,
                 int b)
  {
     return a + b;
  }

Here, the search key which ctags uses for the add function is
"static int add(int a,", taken from the definition. But it's
exactly the same as a the first line of the declaration, and
that is what Vim jumps to for that tag.

A few function declarations in TXR have this issue.

* eval.c (expand_params_rec, do_eval): Make the first line of
the forward declaration different from the first line of the
definition.

* match.c (mf_all): Likewise.

* struct.c (make_struct_type_compat): Likewise.