txr: bugfix: repeat not finding Lisp vars in brace.
The following does not work and is fixed here:

  @(output)
  @(repeat)
  @{lisp-expr ...}
  @(end)

When a Lisp expression occurs in a braced expansion
syntax, it is not not traversed to find variables.

* parser.y (extract_vars): Treat the second element of a
sys:var as a Lisp expression, and find variables. Don't do
this in 128 or older compatibility mode, because then that is
a TXR expression.

* y.tab.c.shipped: Regenerated.