compiler: bump up compiled file version to 5.1.
* share/txr/stdlib/compiler.tl (%tlo-ver%): Bump minor version
from 0 to 1. Because the major version doesn't change, these
files will be loaded by TXR 216 through 243, because those
versions only look at the major number. The upcoming 244 will
continue to do that. What the minor number increase indicates
is that the compiled files do not use the deprecated movi
instruction group. If a future version of TXR removes the
instructions from the VM, it will then insist on only loading
version 5 files if the minor number is at least 1. Version 5.0
files will be rejected, as will version 4 or lower.